added staging + storage user path for sign in

+++ db/seeds.rb
@@ -1,3 +1,5 @@
+return if Rails.env.production?
+
 p "model destruction..."
 Workshop.destroy_all
 User.destroy_all
